python/pyproject.toml
[project]
authors = [
    { name = "Stefano Babini", email = "stefano.babini5@studio.unibo.it" },
]
name = "pystream"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "confluent-kafka>=2.8.0,<3.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "mysql-replication (>=1.0.9,<2.0.0)",
    "mysql-connector-python (>=9.2.0,<10.0.0)",
    "mysql-connector (>=2.2.9,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
